requirements.txt
pandas==2.2.3
matplotlib
streamlit==1.42.1
plotly
scikit-learn==1.3.0
seaborn
numpy==1.23.0
openai
pytesseract
pillow
python-dotenv
xlsxwriter
transformers
torch
datasets
tokenizers
sentencepiece
accelerate
huggingface_hub